nekdo se dotazuje na neexistujici URL jako /software/nastroje/350300996597
git-svn-id: https://zdrojaky.abclinuxu.cz/svn-abclinuxu/trunk@3368 be82d567-0701-4079-a524-1ea639e27e29

+++ src/cz/abclinuxu/utils/Misc.java
@@ -156,7 +156,7 @@ public class Misc {
             throw new InvalidInputException("Řetězec '"+str+"' nemůže být převeden na číslo!");
         if (str.length()!=last)
             str = str.substring(0,last);
-        return Integer.parseInt(str);
+        return (int) Long.parseLong(str);
     }
 
     /**
